switch-config-parse-to-attrs: reorganize the platform module
