+++ src/premiscale/exceptions.py
@@ -1,0 +1,19 @@
+"""
+Custom exceptions for the premiscale package.
+"""
+
+
+from http import HTTPStatus
+
+
+class RateLimitedError(Exception):
+    """
+    Raised when an HTTPStatus.TOO_MANY_REQUESTS code is received.
+    """
+    def __init__(self, message: str ='', delay: float =30.0) -> None:
+        super().__init__(message)
+        self.message = message
+        self.delay = delay
+
+    def __str__(self):
+        return f'RateLimitedError(message="{self.message}", code="{HTTPStatus.TOO_MANY_REQUESTS}", "x-rate-limit-reset={self.delay}")'
